### 2021-06-14 #3364 @whedon
[PRE REVIEW]: OSCILOS_brass: an acoustic solver for brass instruments
**Submitting author:** @AlexanderMacL (<a href="http://orcid.org/0000-0002-5835-216X">Alexander MacLaren</a>)
**Repository:** <a href="https://github.com/MorgansLab/OSCILOS_brass/" target ="_blank">https://github.com/MorgansLab/OSCILOS_brass/</a>
**Version:** v1.0
