### 2021-06-16 #3364 @vissarion
:wave: @astanziola, @krystophny, @pharshalp would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
