### 2021-06-22 #3364 @vissarion
> I don't have MATLAB, so I think that means I can't install/run the software... right?

Unfortunately, yes.
