### 2021-09-16 #3364 @drewbitllama
Hi Vissarion,

I apologize, I just saw this email now—I’d be happy to review the paper if
you still need reviewers.

Cheers,
Drew

On Thu, Sep 16, 2021 at 12:16 AM Vissarion Fisikopoulos <
***@***.***> wrote:

> @AlexanderMacL <https://github.com/AlexanderMacL> as you realize it is
> hard to find a second reviewer (I have also sent several emails) could you
> please recommend a few names to help proceeding with this submission?
>
