### 2021-09-16 #3732 @whedon
[REVIEW]: OSCILOS_brass: an acoustic solver for brass instruments
**Submitting author:** @AlexanderMacL (<a href="http://orcid.org/0000-0002-5835-216X">Alexander MacLaren</a>)
**Repository:** https://github.com/MorgansLab/OSCILOS_brass/
**Branch with paper.md** (empty if default branch):
**Version:** v1.0

- [x] #3752
