### 2021-09-22 #3732 @drewbitllama
Hi all,

Apologies as I'm new to this process. I expressed interest in reviewing a
paper, but I can't find a link to a PDF and the github repo I accepted
seems basically empty :/

Can someone just send me a PDF so I can review the paper?

Thanks!
Drew

On Thu, Sep 16, 2021 at 7:44 AM Vissarion Fisikopoulos <
***@***.***> wrote:

> @whedon <https://github.com/whedon> check references
>
