### 2021-09-22 #3732 @astanziola
Hello @drewbitllama
The pdf is availabe a couple of messages above, in the last message from @whedon

[Here's a link to the pdf](https://github.com/openjournals/joss-papers/blob/joss.03732/joss.03732/10.21105.joss.03732.pdf)
