### 2021-09-22 #3732 @drewbitllama
Thanks Antonio!

I greatly appreciate that, I'll review the paper today.

Cheers,
Drew

On Wed, Sep 22, 2021 at 9:13 AM Antonio Stanziola ***@***.***>
wrote:

> Hello @drewbitllama <https://github.com/drewbitllama>
> The pdf is availabe a couple of messages above, in the last message from
> @whedon <https://github.com/whedon>
>
> Here's a link to the pdf
> <https://github.com/openjournals/joss-papers/blob/joss.03732/joss.03732/10.21105.joss.03732.pdf>
>
