### 2021-09-22 #3732 @astanziola
You're welcome Drew!
That's the first time I review a JOSS paper too, it took a while to get an idea of what to do :smile:
