### 2021-09-22 #3732 @drewbitllama
I appreciate your trailblazing!! I don't review papers too much anymore and
this review process is -definitely- outside of my typical workflow. Thanks
again!

On Wed, Sep 22, 2021 at 9:31 AM Antonio Stanziola ***@***.***>
wrote:

> You're welcome Drew!
> That's the first time I review a JOSS paper too, it took a while to get an
> idea of what to do 😄
>
