### 2021-09-22 MorgansLab/OSCILOS_brass#1 @astanziola
Missing requirements [JOSS review]
_Hello devs! I'm one of the reviewers of your [JOSS paper submission](https://github.com/openjournals/joss-reviews/issues/3732)._

_I will split my comments over several issues, with the hope that it helps addressing each one of them individually. However, I'm new to this JOSS review process, so apologise in advance if this review will be a bit scattered trough time. Also, please let me know if you don't want me to pollute your repository with too many issue, and would rather have a single one with all comments._

---

The JOSS guidelines require the following:
> Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

However it seems to me that your software depends on the `fsolve` function form the MATLAB [Optimization Toolbox](https://uk.mathworks.com/help/optim/ug/fsolve.html), which may not be installed in every system (depending on the MATLAB license).

I think it should be stated as a dependency in your README or documentation.
