### 2021-09-22 #3732 @kthyng
@drewbitllama In case some more information would help — you'll want to look at the github review issue (linked at the bottom of any email you get about this review that comes from github). You can use the checklist at the top of this page to drive your review, opening issues on the [software repository itself](https://github.com/MorgansLab/OSCILOS_brass/) and then checking off boxes as you go (be sure to link to this review in the issues you open so everything is connected). The paper is a part of what you review, but only a small part; most of the review will be running and testing the software itself. Let me or @vissarion know if you have any questions.
