### 2021-09-23 #3732 @vissarion
Thanks @kthyng for answering here!

@drewbitllama In case this is missed in the above discussions, I personally found the links below very useful the first time I did a JOSS review.

https://joss.readthedocs.io/en/latest/review_criteria.html
https://joss.readthedocs.io/en/latest/review_checklist.html
