### 2021-12-21 #3732 @AlexanderMacL
Hi @vissarion and @astanziola

My apologies - I have just realised that the repository submitted with the paper wasn't my personal one, so I've not been seeing notifications. Thanks so much for your help with the review. It looks to me like there's only one issue from astanziola on the repository on MorgansLab (yes, good idea, I'll list fsolve as a dependency in the paper). The issue text implied there would be more issues though. @astanziola could you let me know if I'm missing something? Many thanks
