### 2022-06-14 #3732 @arfon
@AlexanderMacL – my understanding is that @vissarion been waiting for a number of weeks on your input here, and has been in contact over email to try and get your attention.

At this point I would like a re-confirmation from you in the next couple of weeks (by 28th June) that you're still interested in pursuing this publication with JOSS. Thank you!
