### 2022-06-15 #3732 @AlexanderMacL
Hi Arfon, Vissarion,

Many thanks for your email, and my sincere apologies for missing your
previous.

I am very much still interested in publishing this work - as of February
this year it was unclear to me as to whether the reviews were complete,
following Vissarion's email.

It looks like several reviewers were loathe to review MATLAB code, so it
was difficult to get traction on this.

What are your thoughts on a Python port? Or Octave?

Thanks again,
All the best,
Alexander
