### 2022-06-15 #3732 @vissarion
Hi @AlexanderMacL, great that you are still interesting in publishing this work!

Regarding reviewers let me first ping the current reviewers once again.

Regarding the Python port or Octave I do not have experience but if that solution can be used without having MATLAB installed I think it will help a lot to find more reviewers (and will also broaden the user base of your package).
