### 2022-10-17 #3732 @krystophny
> Hi @krystophny @pharshalp you are both experts in acoustics and matlab (that turned out to be a rare combination !), would you like to help us reviewing this submission for JOSS?

Sure! Could take until beginning of November though, depending on free time slots.
