### 2022-10-26 #3732 @vissarion
Hi @augucarv,  @torbenwendt, @GLizaso, @jmrplens you are experts in acoustics and MATLAB, would you like to help us reviewing this submission for JOSS?
