### 2022-10-26 #3732 @augucarv
> Hi @augucarv, @torbenwendt, @GLizaso, @jmrplens you are experts in acoustics and MATLAB, would you like to help us reviewing this submission for JOSS?
>
> We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @vissarion, I'm really short on time and won't be able to review it. Best of luck.
