### 2022-11-24 #3732 @vissarion
Hi @krystophny could you please update us on your review progress? You have to run the following command in order to create you checklist `@editorialbot generate my checklist` and then fill it out during your review.
