### 2022-11-26 #3732 @krystophny
I am sorry that I have to propose to reject the submission in the current state. I am not willing to invest more any more time into this as a reviewer. The reasons are the following.

1) There are only few git commits, and not a single one is from the main author. Furthermore, it derives from another package OSCILOS_long, and it is not clear what code is re-used. I can therefore not verify the authors' contributions.
2) The software is not used in any peer-reviewed paper and seems to be known by very few users judging from the GitHub page.
3) There is no automated testing whatsoever, as the software is GUI-based in MATLAB.
4) The paper is minimalistic, with no background or references to other papers.
